put common antinu_cleaning functions in separate file
